"Added pagination and export functionality to Properties and Units datatables in Owner Dashboard"
This commit adds pagination to the Properties and Units datatables in the Owner Dashboard, allowing for easier navigation of larger datasets. Additionally, a "Download" button has been added to the end of each datatable, allowing the user to export the data as a CSV file. The new functionality is enabled through the use of the PrimeVue DataTable component's `:rows` and `paginator` props, as well as the `exportCSV` method. The `propertiesTable` and `unitsTable` refs have been added to enable access to the methods of the respective DataTable instances.